Move common util scripts to apps from each example folder
util scripts in ssd-object-detector and live-object-detector examples
are same, move them to apps folder, future util functions can be
maintained together.
